+++ src/onix/onix.js
@@ -697,14 +697,7 @@ onix = (function() {
 	 * @member onix
 	 */
 	onix.info = function() {
-		console.log(
-			"OnixJS framework\n" +
-			"2.5.4/20. 5. 2016\n" +
-			"source: https://gitlab.com/LorDOniX/onix\n" +
-			"documentation: https://gitlab.com/LorDOniX/onix/tree/master/docs\n" +
-			"@license MIT\n" +
-			"* - Free for use in both personal and commercial projects\n"
-		);
+		console.log("{ONIX_INFO}");
 	};
 
 	onix.factory("$di", function() {